Implementação de endpoint get Linhas por Parada

+++ backend/ViewModels/Responses/ParadaUpdateResponse.cs
@@ -2,9 +2,9 @@
 {
     public class ParadaUpdateResponse
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
